MTS-7 app start refactor + initial setup with mongomemoryserver for testing
