+++ auth/Dockerfile
@@ -2,7 +2,7 @@ FROM node:alpine
 
 WORKDIR /app
 COPY package.json .
-RUN yarn
+RUN yarn --production
 COPY . .
 
 CMD ["yarn", "start"]
