+++ auth/src/test/setup.ts
@@ -1,0 +1,25 @@
+import { MongoMemoryServer } from 'mongodb-memory-server';
+import mongoose from 'mongoose';
+
+let mongo;
+
+beforeAll(async () => {
+  mongo = await MongoMemoryServer.create();
+  const mongoUri = mongo.getUri();
+
+  await mongoose.connect(mongoUri, {});
+});
+
+beforeEach(async () => {
+  const collections = await mongoose.connection.db.collections();
+
+  collections.forEach(async (collection) => {
+    await collection.deleteMany({});
+  });
+});
+
+afterAll(async () => {
+  if (mongo) await mongo.stop();
+
+  await mongoose.connection.close();
+});
