+++ auth/tsconfig.json
@@ -18,7 +18,7 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "esModuleInterop": true,
-    "types": ["node", "express"],
+    "types": ["node", "express", "jest"],
     "typeRoots": ["./src/types", "./node_modules/@types"],
     "lib": [
         "es2019",
